Fixes issue 38085: Prices in Invoice lines of discounts copied by Copy Lines process are updated to 0 when completing if a price adjustment exists.
The problem is happening when we are copying the invoice line with the basic
discount, in the Copy Lines process and M_PROMOTION_CALCULATE function resets the
price for this line too. As lines with basic discount have zero grosspricestd,
gross_unit_price and line_gross_amount are set to zero too.

We should avoid to reset prices in the M_PROMOTION_CALCULATE function for those
lines related to basic discounts. They are the invoice lines related to an invoice
discount, or those when product is related to a basic discount.

--HG--
extra : rebase_source : 474a1b618bd6ecf3736b96280b492fcf01a91048