Fixes issue 38218: 'Delete failed' error in Reset Accounting process if it is launched for 'Organization' type org without legal children
When the organization has an 'Organization' type and it doesn't has any legal
children, then a not well formed SQL sentence was throwing an exception because was trying
to filter orgs with the ad_org_id column in an empty list of IDs (ad_org_id in ()).

To fix that, if there isn't any organization to be affected by the process, then it is avoided
the execution of several steps of the algorithm, including don't execute queries
that uses the organization id list as filter.

So, if the organization has an 'Organization' type and it doesn't has any legal
children then the logic for the records deletion is skipped.