Issue 38258. Remove 'RELATED_INFO' button aka "Linked Items" as it is obsolete.
That buttons is to show the old 2.50 style 'Linked Items' popup which is only
useful when run inside a (2.50 style) generated window.
However the only prepare*Template methods not already removing the button is
the prepareSimpleExcelToolBarTemplate which is only used in 3 reports (1 in pi
and 2 in modules.
Double checking in 2 of those reports shows that the Linked items button was
present in toolbar but in fact did not work and fails with a missing tabId error
as it is not run from inside a generated window.

So remove the button completely as being obsolete.

--HG--
extra : rebase_source : 58b4843056a4b7ea2c9ce418cc749ba2a371bbeb