Issue 37500. Stop sending ad_module.version_id of module 0 in heartbeat
That is is not useful as only tracking core revision (and not any modules).
As part of its removal stop sending the field in data of heartbeat.

Butler code (HeartbeatServlet.java) is using getStringParameter and setting
value for dal insert into ad_heartbeat. But as that column is nullable should
be fine to push this pi change without change to butler code.

This change also stops calling setCodeRevision for the local insert into
ad_heartbeat_log. That column is mandatory but has a default value of '0'
so should work unmodified.

--HG--
extra : rebase_source : af272454a7b38f852cf7f12393cd2f171fae1f82