Fixed bug 38100: Link to parent trees hides non-active records when filtering.
Fixed by adding a parameter to prevent filtering of non-active records in
tree views. To ensure this parameter is used only in tree grid views, it is
passed as a parameter instead of hard-coding it in the DatasourceService.

--HG--
extra : rebase_source : e074a1e75c207a196761de50454be4364e4c9dc4