Fixed issue 38421: Grid selector dropdown is too small.
Dropdown width calculations doesn't take into account that grid fields length
may be longer than the available viewport, so part of the space left of the
field is not visible in the viewport. ScrollLeft property gives the distance from
the beginning to the point where fields starts to be visible in viewport, so
subtracting this value gives the visible left space available.

Also, the code is refactored to make this calculations more readable.