Fixes issue 38206: Cost Adjustment Process generates too much lines
When adjusting a product used as Raw Material in a Production the same
m_transaction can be added several times in different cost adjustment lines.

Now when exists a non processed cost adjustment line for the same transaccion,
currency, accounting date and source the adjusted amount is added instance of
create a new adjustment line.