[Prepayments] Fixed error that with the payments creation.
There was an error when trying to ger the remaining PSD of an order. As the flush is not already done, when searching for the remaining PDS when the PSD just have been created  by a query, that PSD was not being found.

--HG--
extra : transplant_source : %F5%96R%88%24%08%98/%D4%85p%F2%F8%A7%DC%87Y-%AA%60