Fixed bug 38422: Characteristic description selector not working with filter.
Product characteristics filter assumes the parentGridEntity is always a Product,
but for the Add Product dialog, it is a ProductWithCharacteristics entity. In this
case it is needed to add a reference to a Product instance to make filtering
work as expected.

--HG--
extra : rebase_source : c0f2a06f6ff561d48dce52c205e56d49209fbeab