Fixes issue 38467: Fix precision of payment amount read from json in the order loader.
Payment amount when read from json was passed to Double and then to BigDecimal which produce precision problems.
The precision problems makes that the check if there is change to be true, when in really there isn't change,
which produce the behaviour reported in the issue.

Now readed as string from json and passed directly to BigDecimal.

--HG--
extra : transplant_source : %25%23%A6xiz%E8%0E%8F%25WgZG%3E%ED%EE%02%C9%CF