Related to issue 38422: Characteristic desc. selector not working with filter.
Adding filtering by ProductWithCharacteristicValue when no filter is applied in
the parent grid. This makes the characteristics selector behavior coherent with
the rest of selectors: if no filtering is performed, shows only the
characteristics present in the grid entries instead of all available values,
same as if we were filtering.

--HG--
extra : rebase_source : 288324972709e99d5fdd1bbfbdfb1be33213c235