[Prepayments] Performed the payment generation in some special cases.
In the cases in which the generating payment was of different sign to the order sign, for example when generating a payment of amount 10 for a negative ticket, instead of generating a payment with a unique detail of 10, there was a payment with two details, one of 20 and the other one of -10, that had a sum equal to the first case. Anyway, the flow has been performed to have the expected result.

--HG--
extra : transplant_source : E%AAn%EF%FD%0F%94%BD8x%EC%3F%CA%99E%F1%CEv%A8%85