Fixed issue 38573: The save action of the ticket is being executed more times that should
When doing the save action in a calculateReceipt function, now the save is only done if there are no pending calculateReceipt processes.
When modifying the line quantity, the save action is no directly called now, is called from the calculateReceipt action that is executed when setting a new quantity.
In the delete lines process, now the save action is not executed when doing the calculateReceipt, because is called after that. In this flow, is necessary to call it later. Also, to finish the delete lines process the Web POS was waiting for the save action, now the action is launched but the user can continue without waiting for the response.
The calculateGrossAndSave funtion is sent with the true parameter when loading a receipt, otherwise the ticket is not correctly stored in the Web SQL, causing problems if an user presses the F5 having the ticket loaded.

--HG--
extra : rebase_source : e19eb6734d1ef63f37e133d8cab4807d1c32e8ec