[Product Locked/Discontinued] Fixed errors.
The stock is verified when trying to navigate to pay open tickets.
The 'producAssortmentStatus' property is changed to 'productAssortmentStatus'.
When there's no a warehouse configured for a product, the stock is set as 0.
When a stock request is returned with a connection problem, it is managed from the callback error popup.