[clfr] Refactorized the way the orderedQuantity and the operativeQuantity is returned.
If the processin line is an Order Line and it is related with a Shipment/Receipt
line then the ordered and operative quantities used will the movement quantity of
the Shipment/Receipt Line. In other hands will be the ordered/movement quantity
showed in the P&E.

Added some logs to the CreateLinesFromUtil.