[clfr] Fixed the Order Line updation at invoice header level.
If exists more than one order related with the invoice lines then the Order header
will be null, in other hand it will be the order related with invoice lines.

Was necessary to call a flush after each new invoice line was copied to be able to
get the information of the order lines related to them.