[clfr] Fixed query to get the shipment/receipt lines related to an order line
CreateLinesFromUtil.getRelatedShipmentLines method was improved to correctly filter
the shipment in out lines related to an order line. Were added more restrictions
according the order line belongs to a sales or a purchase order, and a new relation
with the procurementPOInvoiceMatchList of the in out line was added in case of Purchase flows.

This method is now returning only the required information to later be added to
its JSON representation. This includes get the ordered quantity correctly calculated.
It returns a list of ShipmentInOutLineData objects, created to store specifically
this required information to store later in the JSON and be used by the process,
and nothing more.