Now it's possible to convert a receipt that is being replaced to a layaway.
Befor this change, it was not possible to convert the new receipt generated during a C&R to a layaway. Now, when the ticket is still in draft status, if the ticket doesn't contain any fully or partially delivered line, it's possible to convert the receipt to a layaway.

--HG--
extra : transplant_source : %EF%2C%28%E3%9E%87K%D1Y%03%F8%F7%2C%7D%60-%2B%BB%AE%89