[clfr] Removed the information message in the Create Lines From InOut P&Es
This information message showed in the InOUt P&E is not needed because the information
is shown in the lines is the same that will be copied to the invoice line when
it is created. It is not the case of Create Lines From Order due the possibility
of the order lines to be related to shipment/receipts lines.