[clfr] Avoid exceptions when filtering by ordered/movement quantity field
As the Ordered/Movement quantity fields are usually calculated by the substraction
of the matched/invoiced/shipped quantities to the ordered/movement quantity it is
needed to retrieve it by the sum of related lines.

When the P&E is filtered by the Ordered/Movement Quantity, the platform takes the
column definition (orderedQuantity - SUM(...)) and adds it to the WHERE CLAUSE of the
generated query, adding the entered filters in the UI to correctly filtering all records
by the column.

The way to avoid the exception is to replace the aggregation function (sum) by a subquery
recalculating the same value representing the matched/invoiced/shipped totals.