[clfr] Improved the way the AUM is retrieved in the PEs.
If AUM preference is enabled then the AUM used in the line is retrieved else the
default AUM. If the AUM preference isn't enabled then it is not needed to get this
information because it will be not used in the processing logic and it is returned
an empty string instead of execute the logic when AUM is enabled.

Also it avoids DIVISION BY ZERO if qtyOrdered or movementQuantity is ZERO at
division time when orderQuantity is calculated.