Fixed issue 38099: Upgrade from <3.0PR15.4 preserves old pool configuration.
3.0PR15.4 includes the Apache JDBC connection pool module as part of Openbravo
distribution and is set as the default DB pool. This changeset enforces upgrading
to the new pool instead of preserving the old Apache commons pool.
This implementation has been tested in the following scenarios:
- From an 3.0PR15Q3 instance with the parameter db.externalPoolClassName not set
or empty: After the upgrade the parameter should refer to the bundled Apache JDBC
pool module.
- From an 3.0PR15Q3 instance with the parameter db.externalPoolClassName set:
Property db.externalPoolClassName should be left as-is.
- Upgrading from a newer version (e.g. 3.0PR16Q1): This validation should not be
executed.