Fixes Issue 38735: incorrect query in ReservationUtils
HQL query in getReservationStockFromStorageDetail method
in ReservationUtils.java had ")" end brace at end with out
"(" start brace in the query. This is not correct but was
allowed in hibernate version 3.6, but hibernate version 5.3
does not allow such error in query. Hence the query is modified
by removing the ")" end brace.