[clfr] Avoid join with M_PriceList table to improve performance in CLF PO
Avoid the join with M_PriceList table because it is only made to get if it includes
taxes or not. As it can be retrieved from the order header, it is better to get from it
instead of do a join with the M_PriceList Table.

Also fixed the copyright in OrderLinePEHQLTransformer.java

--HG--
extra : rebase_source : 0c325eb87fa1ddea7cd05cbb434fa9ce014c9e2c