[clfr] Disable sorting for UOM, Product, Warehouse and Storage Bin fields.
To improve the performance, these fields are retrieved by a subquery instead of
doing JOIN with the tables. As this left clause is used by platform to build the
query is used when filtering, it fails when this subquery is used as ORDERING condition
of the generated query ordering by the selected column.

To keep the improvements introduced by the subquery, user will be able
to filter by the column but not to order by them.