[clfr] Get quantities directly from the JSON instead of the order/inout line
As the selected lines are pre-processed before creating the new invoice lines,
the JSON containing the lines to be copied already has all the quantities calculated.
If it is an order line without any shipment related or it is an inout line then the
quantities are the same that are shown in the PE. In case the order line has inout
related, then the process add as many inout lines as the order line have related,
and these new JSON entries are going with the quantities calculated.
In both cases it is not needed to take thhis information from the objects, just from the
JSON line directly.
Removed unneded methods from CreateLinesFromUtil, and move to package the used ones.