Fixes bug 38801: Error in TriggerHandler.enable doesnt leave triggers disabled
The problem was that if the flush indie TriggerHandler.enable (invoked when retrieving DAL's connection) failed, the clear method was not
invoked, and the ThreadLocal was not reset. The next time the thread was reused, it would work with disabled triggers.

Now that doesn't happen anymore because the flush is done inside the try block.

--HG--
extra : rebase_source : ed89da35fb557175e5975e6db80d7ea39290dc57