[clfr] Improved the Create Lines From Purchase Orders process
It is made a LEFT OUTER JOIN with the invoice lines of the invoice that is processed
related to the current order line.
Was removed the NOT EXISTS subquery used to exclude all the order lines related to the
processing invoice from the where clause.
In the GROUP BY clause is subtracted the invoiced quantity, this way it is not showed those
lines wich the subtraction is different than 0