[clfr] Some changes to calculate quantities
To improve performance and make more legible the resulting query in CLFR PO, was
replaced the subquery used to get the already invoiced quantity in the processing
invoice by a direct SUM() clause.

This change causes that the generated SQL query appends to the query an aggregation
function (SUM) to the main query when the PE is filtered by Ordered Qty or AUM Qty.
Because of it, these columns as marked as NOT FILTERABLE.

Also was changed the methods in transformers to always return the operative quantity
from the formula calculating the ordered quantity.