Fixed issue 38825: Ensure that an order is going to be saved before setting the id
In Web POS, when doing a save action, if the order doesn't have an ID, the ID is
generated and then the order is saved. But there can be the situation in which
the ID is set and the order is not saved, so that order won't be saved again.

Now, before setting the ID, it is verified that the order will be saved.