[clfr] Improved performance in CLF SO process.
Avoid the join with C_ORDER table, and now is doing the filtering by the order
fields directly from the view instead of the C_ORDER table.
Was needed to add all the order fields to the C_INVOICE_CANDIDATE_V view, and also
those fields used in the select clause to show the order information in the PE.

Was adapted the OrderLinePEHQLTransformer to get the information from the view (ic)
when the process is executed in the Sales Invoice, and to get the information from
the order when it is executed from the Purchase Window.