Fixes Issue 0038667: "Not Balance" error received when posting a reconciliation (although suspense balancing is not needed)
getBalance() method should return ZERO as transaction lines
in reconciliation are always balanced.

However transaction line linked with payment are skipped for
posting if payment method & financial account in payment
is not configured with upon reconciliation use account
in payment in and payment out allowed section in payment
method tab of financial account.