[clfr] Removed the order/inout identifier column from Create Lines From PEs
Columns used to build the identifiers have been included as independent columns in
the Create Lines From PEs, because of it it is not needed the identifier columns
as it will be a redundant information.

This change includes remove the column from the select clause of each HQL Table,
and also from the AD. salesOrder and shipmentReceipt fields are removed too, and
was clean the InOutLinePEHQLTransformer to remove the code replacing the business
partner name from the InOut identifiers building.