Fixed issue 37460: Check the delivered status to open receipts.
When searching for an order, now it delivered status is checked in order to be able to load the ticket. If the (paid) order is not delivered it won't be able to be loaded. This affects to the canceled orders. A new index is added in the delivered column to have faster queries.

Also, the loaded canceled tickets has the order status flag set to 'Canceled', and there's also a new label for the lines which are not totally delivered to be shown as canceled.

Finally, the delivered status is checked for tickets created during a C&R process.

--HG--
extra : rebase_source : 500a7bf47f2a5a869324acce7441b71af8785996