Fixed issue 38845: The related receipts of an order are not being loaded correctly
Now, to remove a record from the this.selectedOrders array (the array that stores the orders to load) a 'for' loop is used instead of the 'indexOf' function, that was not working correctly. Also, when adding again an order to the array, instead of introducing the full backbone model, only the json with the properties is stored, to maintain the consistency.