[Verified Returns Refactor] Change the way to create a component so it can be used in the tests
It was failing that when opening the verified returns popup and later the open receipts popup,
some elements like the lens button don't get id because is duplicated the element,
this is because is reusing the same element.

Changed to add this element on runtime with diferent name so both element has id for the tests.