[Cancel Layaway Refactor] Refactored the CL and VL processes
The VL process is done only for those layaways without any payment. Now when clicking on the 'Void this Layaway' menu entry, the process is executed directly not being needed to complete the ticket as before. Also, a voided layaway has now the canceled flag.
In the CL process, the negative ticket is generated now in the Web POS. The lines and the taxes are calculated there. After that, the ticket is generated in backend using the methods in the OrderLoader class, extending it in the CancelLayawayLoader.
When doing F5 in a CL or a C&R ticket, now the tickets are correctly restored.
The amounts of the CL and C&R tickets with a sign different to its real negative status are now managed correctly.

--HG--
extra : transplant_source : g%9F%D4%136%A6%AB%5EK%D0%3A%BAg%FA%01%F16%B0j%AB