Fixes issue 0038909:"PLM Status" field in Product window should be renamed
As part of the project Product Lifecycle Management a new field has been added
to the Product window to set the status of the product. The label for that field
is set as "PLM Status", a name that can be improved.

The PLM Status field was renamed to Lifecycle Status in Product window.