Fixed issue 38922: The variants of a product are not shown.
Currently the manage variants datasource is limited to 1000 variants, and
if this limit is exceeded no data is calculated even if performing filters.
This process needs further optimization and refactor.
This changeset adds the possibility of showing an error message when a exception
in the datasource is thrown. This let us to show a different message when
variants limit is exceeded or no variants are available.