Related with issue 38903: Backs out changeset e17a08d06f29, change too central
The backed out changeset could have an impact on performance. It changed the initialization of
OBContext, which is done many times, and a small impact on performance can add up.

It will be replaced by a fix that will address the specific event handler where the issue was
being reproduced.

--HG--
extra : rebase_source : 4f086358114fbd0d4bc3791cdec777efd860b683