[Verified Returns Refactor] Fix to not use isDelivered
In the first implementation of the refactor of verified returns selector,
we use the variable isDelivered, thinking that it means that the order
has any line delivered.

After, we notice that isDelivered means that the order has been completely delivered.

For that reason, changed the query for verified returns selector to be the same that
open receipts when you select order type.

--HG--
extra : rebase_source : ae0e7a4be0f71c1e1088223e3549748854d38e36