Fixes issue 39190 : fix the grouping for invoices generated in cashup process
Add ad_user_id to the grouping in missing sqls.

Add condition that ad_user of the invoice and the order is the same.

If ad_user null, used coalesce to set to '1', for compare since nulls can not be compared.
Note: Used '1' instead '', because oracle consired '' like a null and con not be compared neither.

--HG--
extra : transplant_source : h%87%F3%DEq%07%60%29%3D%26pg%BC%009%DA%A2f%92%99