Fixes issue 39191 : fix the grouping for invoices generated in cashup process
Add ad_user_id to the grouping in missing sqls.

Add condition that ad_user of the invoice and the order is the same.

If ad_user null, used coalesce to set to '1', for compare since nulls can not be compared.
Note: Used '1' instead '', because oracle consired '' like a null and con not be compared neither.

--HG--
extra : rebase_source : 55c1c874e6d946f3367e6f721a50c3339aec9770