Fixes issue 39168: A payment can be reversed more than once with slow connection
Was possible to reverse a payment more than once with slow network connection.
It was caused because could be a time between the payment is reversed and the button
is hidden where it is possible to execute the process again. As the process of
reversing the payment happens in the server-side and the UI delay happens in the
Front-End the Reverse Payment button is hidden before setting up the parameters
so when the process is executed the second time the payment is already reversed
and linked to the reversed payment.
To avoid this bug, it is validated that the processing payment is not already reversed.
In case it is, then an error message is thrown explaining it is reversed.

--HG--
extra : rebase_source : e989672fc41632b44e271e529dbc6011bf11cbeb