Fixed issue 39201: Menu overlapped by messageBar if receiving a response from a closed window.
Instead of dissapearing, the menu were overlapped with a error messageBar which
originally belonged to the window that made the request (and were closed before
it finishes). This fix ensures the datasource window is not currently destroyed
to show the error message.

--HG--
extra : amend_source : 7c9d4cc356c248046e4e36f9fd9138ba323c7b81