Fixed issue 38159: Memory leak in referenced inventory unbox process
The leak is caused by an invalid call of fireOnPause method on selector
with a targetRecordId (when clicking from the looking glass icon of a field).
This call gives an invalid delay value that causes the timeout to be called
indefinitely.

--HG--
extra : amend_source : 20f33f5c9b5ff7a5a39fdf9c577ec69827efdf61