Fixes issue 39316: Unneeded error shown in log if origin is empty
The problem was that sometimes the Origin request param is set to "null" (string with "null" as
content). In that case, the null checks were not met, and an actual check against the allowed
origins was done. This check failed and the error was shown in the log.

If the origin is empty, null or "null" the check does not need to be done.