Fixed bug 39391: Movements in same bin for non-attribute product
The goods movement created by AWO when a product has no attribute are automatically set with m_attributesetinstance_id = '0'.
In a normal movement the m_attributesetinstanceTO_id is usually null (unless an attribute change or a RI related flow involved). Thus the comparation has '0'='X' which avoided to raise the expected error.

Note that in Core the m_attributesetinstance_id for non-attribute product is NULL, so the issue is not reproducible in Core. Besides the UI controls this scenario very well.

The solution is done in the trigger rather than in AWO because it is safer and easier to develop.
It runs the locators check only when m_attributesetinstanceTO_id is null.
If m_attributesetinstanceTO_id is not null it implies an attribute change or a RI related movement, and in any of these cases it is allowed to run the movement within the same bin.