Fixes issue 39412: Prevents unintended formitem value change if focus changes
There was a problem in this redraw function [1]. It did the following:
1) Get the value of the focused item (this.getFocusItem().getValue())
2) Redraw the form (this.Super('redraw', arguments);)
3) Restore the value of the focused item if the value changed during the redraw (this.getFocusItem().setElementValue(focusItemValue);)

The problem was that under some circumstances the focusedItem itself changed during the redraw, so the form items returned by this.getFocusItem()
in steps 1) and 3) are different, and the value from one form item was copied into another form item.

This problem has been fixed by storing in a variable the focused form item before the step 1), and the use that reference in the rest of the
function instead of reinvoking this.getFocusItem()

[1] https://code.openbravo.com/erp/devel/pi/file/tip/modules/org.openbravo.client.application/web/org.openbravo.client.application/js/form/ob-view-form.js#l2157 [^]

---
M modules/org.openbravo.client.application/web/org.openbravo.client.application/js/form/ob-view-form.js: Enter commit message.  Lines beginning with 'HG:' are removed.