[Delivery Conditions] Fixed error that used to deliver deleted lines
It was possible to generate shipments for lines that were delivered. If a synchronization process fails after the hook that sets the quantity to deliver but before the synchronization process is sent to backend, the order is not restored to the initial state, and the quantities to deliver continues being set to the updated quantities. Then, if the user deletes some line and synchronizes again (without errors), the deleted lines have a quantity to deliver set, so shipments are created for those cases.