Fixes issue 39405: Tab display logic works when direct opening a header tab
Before this change, when a window was opened by clicking on a column title, the display logic of the
tabs of the window were not evaluated. This was done because if the tab opened was a subtab, that tab
should be visible, regardless of that tab's display logic.

But there is no need to enforce that logic when the tab opened is a header tab. This changeset evaluates
the display logics of the tabs if the tab opened directly is a header tab.

--HG--
extra : rebase_source : 81d4d17a9a99ecc884ee7bac158f1bee3a96dd06