Fixes issue 39530: No stock error for Orphan lines in Return to Vendor Shipment
By default the UI will show an error if there is no stock available and there
not exists at least one storage bin linked to an inventory status configured as
overissue = Y inside the Return To Vendor Shipment's Warehouse.

If there is no overissue bin inside the RTVS warehouse, then the No Stock error
is shown when the line is selected. Also when the RTVS line is inserted, the storage
bin is automatically filled with the overissue bin and when the RTVS is processed,
the stock goes to negative in that bin.