[Omnichannel] Fixed error when generating invoices
To create an invoice line, the current shipment lines are searched for, to generate a line for each shipment line. The problem was that even when the shipment lines have been created, as the flush haven't been done, those lines cannot be searched for using DAL. They must be taken from the hibernate model.