Related to issue 39586: It is not possible to apply a manual promotion/discount to a negative line
Change message to indicate current promotion could not be applied to negative
lines.

Use new flag definition to allow or to deny apply a manual promotion to a
negative line.

Save flag value in promotion applied.

If the user change the quantity of the line using Return Line button or Return
this receipt, even if the manual promotion is marked to be applied in negative
lines, the discount disappears.

The problem now is fixed for both cases.