[Omnichannel] Removed check before doing a cancel and new
There was being checked if a line is able to be alone in a new receipt, to avoid to remove the line later with the quantity rules.
As is not possible to cancel only a service (because the service delivery is linked to the product delivery), the check is not necessary.