[Omnichannel] Updated the adjustPayment and getPaymentStatus processes
Now is the adjustPayment the process that checks the negative status, not the getPaymentStatus. The getPaymentStatus process only reads the value set during the adjustPayment.
Also removed calls to the adjustPayment and save processes that were being done twice for a single process (when a calculateReceipt is called it must be the one which calls directly to the adjustPayment process).
The isNegative function also checks directly the negative status without calling the getPaymentStatus process.