Fixed issue 39676: Error when calculating the taxes in the "Copy from Orders" process.
Executing the "Copy from Order" process, the bill partner location was calculated
in the UpdateTax.getMaxBusinessPartnerLocationId() method, taking the maximum value
of the bussiness Partner? locations.

Now it will use the Bill Location of the order if defined or the last partner address
in other case, similar to the flow when the line is created manually and the SL_Order_Product
callout is called.

--HG--
extra : rebase_source : 73849271edce0acb9faf0726c54b423add4e01b0