[Omnichannel] Change when generating the invoice
When loading the PSDs that are not assigned to any invoice, they are now always ordered to avoid to have different results in oracle and postgres. This was causing that an invoice could be paid by different payments depending on the database that was being used (even if it was being done correctly by any way).