Fixes Issue 0039710: Landed cost can be reactivated TWICE
** Inorder to avoid process or reactivate twice, processing flag
is used for locking mechanism. At the beginning check is done
whether the flag is checked, if yes error is shown else flag is
set as Yes and process or reactivate action proceeds. At the end
flag is reset as No.
** Checking of processed flag or document status (Draft) is also
done while reactivation process whether the landed cost is
reactivated by other process.