Fixes Issue 39822. Fixes issues when merging new cost adjustment lines with existing ones.
To avoid creating more Adjusment Lines than needed, the process
tries to find existing cost adjustment lines and it adds the cost
adjusted to an existing line if any.
But when looking for existing lines it was not taking into account
the unit cost, negative correction or backdated flags.

To do so, it is necessary to send the parameters to the methods
that creates the lines. But adding more parameters to this
functions was too complex, as they already had many parameters.

What has been done is to create a new class to store the line information.
This class is initialized with just three parameters and the
rest are set to the default values.

Then, the developer can change any parameter with the setter methods.
This makes the code more clear and clean, instead of having a
function with more than 10 parameters, being 4 of them booleans,
now it creates a new object with three parameters and then it
sets the boolean values with methods that have a name with more
meaning.

After this change, all the information of the line is set through this
new object, so there is no need to change the line object afterwards,
since all the different parameters are set to the LineParameter object.

Changes:

Database:
* Change default value of needsposting column to use the same
  one as in AD and avoid confusion

CostAdjustmentLineParameters:
* New class to store lines information
* Does not allow to create adjustment lines with empty cost adjusment
  or empty Transaction
* Does not create an adjustment that is both negative and unit cost.
  This was not controlled before

CostAdjustmentUtils:
* Deprecated old methods that does not send the new flags
* Created new API to use the new LineParams object
* Modify query to take into account new flags
* Modify insert method to set all the parameters for the new
  Cost Adjusment Line

CostingAlgorithmAdjustmentImp:
* Modify inser line method to use new API

Adapted all the calls to old methods to use new API

--HG--
extra : rebase_source : 84664df38a6c355168e53e125efec392b50ecfac