[Omnichannel] Fixed issues related to omnichannel
The loaded of a C&R was not being checked correctly. The id of the invoice line taxes were being set with the same id of the order line, so in the case when 2 invoices were being created for a line (because a service can be related to different lines, and can be delivered and invoiced in multiple synchronizations) and the id was being duplicated for the same table.
When a line was not being delivered could be invoiced (wrong) with the 'after delivery' invoice term.
When invoicing a line that was previously partially invoiced, the invoiced quantity was not being correctly set (the previously invoiced lines were being invoiced again).
When printing an invoice in 'Pay Open Tickets'.
Loaded the delivered quantity of the deferred services, and used to set the quantities to deliver.

--HG--
extra : rebase_source : 3bb1e1284b0dc1b834ac42382b164f17194b2f41