[Omnichannel] Changes when creating payments in OrderLoader/InvoiceUtils
Now, instead of create the payments during the order loader and then distribute during the invoice utils, the payment schedule details without any payment detail that are not assigned to an invoice are set to the invoice. In that way, when creating the new payments during the orderloader are directly created with an invoice set.

--HG--
extra : rebase_source : e2ea6f4f64b6d8cfcd64e5f75c8d08dc981d4048