Fixes issue 39781: It isn't possible to process shipments from a layaway line
When creating a Goods Shipment lines  by the Create Lines From process, it was possible to
include layaway orderline. It should not be possible to process shipments from a layaway line.

To fix this, was added a new extension point to the obpos_deliver_layaway to raise an error
when was created a shipment from a layaway line.