Fixed issue 39872 : Receipt should be removed from database even if orderList is empty
* With Sessions module, Cashup window will be loaded before rendering the pointofsale window whem sessions closed in backoffice.
In that case, if receipt is saved in database and not completed, It will shown in cashup window to delete it.
OB.MobileApp.model.orderList will not be initiated while deleting the receipt.
So if orderList is present, receipt to be removed will be deleted from database and orderlist
else receipt will be removed from database only