Fixed issue 39952: Problem with the invoice line-shipment lines relations
Now when a shipment is created after the invoice is created, the relation between the invoice line and the shipment line is set.
Also now in the InvoiceUtils.java class it is supported to create an invoice for a line that has already an invoice and a shipment, but the invoiced quantity is lower than the shipped quantity. This situation is cannot occur in the Web POS at now.